requirements.txt
streamlit>=1.32
pandas
plotly
pgeocode
openpyxl
python-dotenv

